+++ collective/signupsheet/browsers/resources/signupsheet.js
@@ -1,0 +1,9 @@
+/**
+ * Script for collective.signupshet
+ */
+
+(function($) {
+	$(document).ready(function() {
+		
+	});
+})(jQuery);