+++ docs/HISTORY.rst
@@ -7,6 +7,11 @@ Changelog
 - Removed some bad tab condition expressions [keul]
 - Fixed security issues: some views were not protected
   with right permissions [keul]
+- Merged features from `Products.SignupSheetNotification`__
+  (an unreleased add-on for original SignupSheet package) [keul]
+
+__ https://svn.plone.org/svn/collective/Products.SignupSheetNotification/trunk/
+
 
 0.1.2 (2014-09-19)
 ------------------